Add a flag to manually specify the waif type of a v4-style database.
So, for instance, if your v4 database has waif as type 10, but Stunt has waif as type 13, you can specify `-w 10` and Stunt will allow type 10 to be read from the database as waifs. It will then automatically handle conversion when the database is next saved. The -w flag is only needed for the first run of a waif-enabled Stunt server to convert the database.