Add an experimental ancestor cache.
In profiling, I found that a lot of time was spent searching for ancestors. In fact, every property lookup.
By caching ancestors, we're able to significantly speed up loops that rely on property lookups. However,
since the repercussions of doing this are unknown, there's also an option.h flag to disable it.