Bandaid patch:
Somewhere in the server, we're losing more references to emptylist than we are gaining references. This results in the refcount hitting 0
and the memory getting freed. The next time new_list(0) gets called, we either fail an assert (if garbage collection is enabled) or we
just crash because we're accessing memory that's already been freed. This bandaid effectively freezes the refcount and makes emptylist
a static Var that never gets freed.