Add a yin() (yield_if_needed) function.
Analogous to the suspend_if_needed function in LambdaCore databases,
this function will suspend the currently running verb if it finds
it's approaching a tick / second limit. Optional arguments allow you
to specify the time to suspend, the minimum amount of ticks left,
and the minimum amount of seconds left. This way you can replace both
suspend_if_needed and manual lines of the form: ticks_left() < 2000 &&
suspend(0).