Fix a crash in connection_name_lookup.
Maybe. It was possible that somebody could disconnect before the lookup
completed, resulting in the network handle being passed to the rewrite
function being freed. So now, instead of passing a pointer, as pass an
objid and look up the network handle in the callback function. That way
if the handle has already been freed, we just back out.

In theory, though, it's probably possible that the main thread could
still free the network handle after we've looked it up in the callback
thread. To fix this properly, I think we'll actually have to keep track
of how many pending connection rewrites there are on the network handle
and change how close_nhandle works. So instead of immediately closing,
it will check how many pending rewrites there are and defer the actual
freeing to the server's main loop, similar to how garbage collection is
handled.

Anyway, we'll see!