Rollback part of commit 01630a.
Objects no longer have the recycled flag set on them when a recycle() begins. Anonymous objects continue to do this for purposes of garbage collection.

This resolves a bug where long-lived tasks called by recycle() (primarily exitfunc) could result in an object being flagged as recycled but still exist indefinitely in an un-recycled state if the task in question was aborted unceremoniously.

If you have any of these objects currently in your database, running recycle on them will get rid of them as it should have.