Reference count nhandles to resolve a potential race condition.
It was possible that a thread could be in the middle of rewriting a
connection_name when the main thread decided to close the nhandle. This
would, of course, cause a crash. So the simplest solution I could think
of was to handle refcounting in the main thread. The main loop will also
delay processing server handles until the nhandle pointed to by the server
handle has an appropriate refcount.