Add additional wizardly debugging information to queue_info().
Now the entire state of a task queue is returned (for wizards) when
invoking queue_info() with an argument. This is certainly not useful for
most people, but for server developers or people debugging very specific
MOO code issues, it can be helpful.

One example would be holding input on a connection and calling
boot_player when it has queued input. That task queue will hang around
forever until somebody calls read() on it or calls flush_input(). With
the new queue_info() information, you'd be able to diagnose that this
was happening by seeing the task queue is not connected but has pending
input and do something appropriate with it.