Handle user-defined signals within the database.
User-defined signals (SIGUSR1 and SIGUSR2) will now be passed to #0:handle_signal as strings. If this verb does not exist, or the verb returns a non-true value, the default server behavior will be assumed:

	- SIGUSR1 forces the logfile to be re-opened.
	- SIGUSR2 forces a server checkpoint.

If #0:handle_signal returns a true value, the server will conclude that the signal has been appropriately handled and not go any further.

This is useful, when, say, you would like a script to trigger a database-level event, but don't wish to use a network connection from outside to notify the database that something needs to happen.