Fix handling of network buffer overflows with TLS connections.
If SSL_write() fails with SSL_ERROR_WANT_WRITE, it will insist on
receiving the exact same buffer as when the error occurred. The MOO's
behavior, however, is to drop data from the beginning of the buffer when
it starts to accumulate beyond MAX_QUEUED_OUTPUT. Since we lose the
previous text_block, SSL_write() gets angry and ultimately the
connection is dropped.

This commit changes the MOO's behavior when it comes to queuing output.
After a connection receives an SSL_ERROR_WANT_WRITE, it sets a flag on
the nhandle. If this flag is set, the MOO will drop output from the
second line onward rather than the first line onward.

Fixes #36