Create a fake stack list if the stack has already been unwound.
It's entirely possible that, by the time our lag check occurs, the
interpreter has already unwound the stack. Rather than store a copy of
the stack list for each frame (which would be needlessly expensive), the
server will now create a fake stack list that contains only the verb
name and the originating object. Prior to this, these verbs would be
ignored completely and return an empty list to handle_lagging_task, so
even with its shortcomings, this is a better solution.

This solves the issue of single verbs not being reported as lagging at
all while also preventing the crash that commit 9139059 addressed.